move BvhNode code into separate file
