+++ src/bvh/mod.rs
@@ -4,8 +4,10 @@
 //!
 
 mod bvh_impl;
+mod bvh_node;
 mod iter;
 mod optimization;
 
 pub use self::bvh_impl::*;
+pub use self::bvh_node::*;
 pub use self::iter::*;
